requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2==2.8.6
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1

plotly~=5.1.0
pandas~=1.3.1
celery~=5.1.2
requests~=2.26.0